refactor: security 상수 분리
