+++ src/main/java/com/part4/team09/otboo/module/common/security/CustomUserPrincipal.java
@@ -1,5 +1,6 @@
 package com.part4.team09.otboo.module.common.security;
 
+import com.part4.team09.otboo.module.common.security.constants.LoginType;
 import com.part4.team09.otboo.module.domain.auth.dto.AuthUserDto;
 import java.util.Collection;
 import java.util.UUID;
