--- src/main/java/com/part4/team09/otboo/module/common/security/LoginType.java
@@ -1,5 +1,0 @@
-package com.part4.team09.otboo.module.common.security;
-
-public enum LoginType {
-  BASIC, OAUTH
-}
