+++ src/main/java/com/part4/team09/otboo/module/common/security/constants/LoginType.java
@@ -1,0 +1,5 @@
+package com.part4.team09.otboo.module.common.security.constants;
+
+public enum LoginType {
+  BASIC, OAUTH
+}
