+++ src/main/java/com/part4/team09/otboo/module/common/security/handler/CustomLogoutHandler.java
@@ -1,6 +1,6 @@
 package com.part4.team09.otboo.module.common.security.handler;
 
-import com.part4.team09.otboo.module.common.security.AuthCookieNames;
+import com.part4.team09.otboo.module.common.security.constants.AuthCookieNames;
 import com.part4.team09.otboo.module.common.security.jwt.JwtTokenProvider;
 import com.part4.team09.otboo.module.common.util.CookieUtil;
 import com.part4.team09.otboo.module.domain.notification.sse.SseService;
