+++ src/main/java/com/part4/team09/otboo/module/common/security/oauth/CustomOAuth2User.java
@@ -1,7 +1,7 @@
 package com.part4.team09.otboo.module.common.security.oauth;
 
 import com.part4.team09.otboo.module.common.security.CustomUserPrincipal;
-import com.part4.team09.otboo.module.common.security.LoginType;
+import com.part4.team09.otboo.module.common.security.constants.LoginType;
 import com.part4.team09.otboo.module.domain.auth.dto.AuthUserDto;
 import java.util.Collection;
 import java.util.List;
