+++ src/main/java/com/part4/team09/otboo/module/domain/user/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.part4.team09.otboo.module.domain.user.controller;
 
 import com.part4.team09.otboo.module.common.enums.SortDirection;
-import com.part4.team09.otboo.module.common.security.AuthCookieNames;
+import com.part4.team09.otboo.module.common.security.constants.AuthCookieNames;
 import com.part4.team09.otboo.module.common.security.jwt.GeneratedToken;
 import com.part4.team09.otboo.module.common.util.CookieUtil;
 import com.part4.team09.otboo.module.domain.auth.exception.MissingRefreshTokenException;
